# GG-8536 Need to support GridPortable in CPP client

+++ modules/clients/cpp/main/src/impl/gridclientprojectionimpl.cpp
@@ -165,7 +165,7 @@ bool GridClientProjectionImpl::processClosure(TGridClientNodePtr node, ClientPro
         }
     }
     else { //no routing
-        int numAddrs = addrs.size();
+        size_t numAddrs = addrs.size();
 
         for (size_t addrIdx = 0; addrIdx < numAddrs; ++addrIdx) {
             try {
